# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Updated workspace profile
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  event:
                    type: object
        "403":
          description: Workspace manager permission required
    delete:
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "204":
          description: Workspace metadata deleted; queued upload-object cleanup may continue asynchronously
        "403":
          description: Workspace owner permission required
...
  /api/channels/{channel_id}:
    patch:
      operationId: updateChannel
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Updated channel
  /api/channels/{channel_id}/workspace-notes:
    patch:
      operationId: updateCodeWorkspaceNotes
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCodeWorkspaceNotesRequest"
      responses:
        "200":
          description: Updated freehand code workspace plan and goal
        "400":
          description: Invalid payload or channel presentation
        "401":
          description: Authentication required
        "403":
          description: Workspace owner or owner-scoped bot permission required
...
  /api/messages/{message_id}:
    get:
      operationId: getMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Message visible to the current user
    patch:
      operationId: updateMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "200":
          description: Updated message
    delete:
      operationId: deleteMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Soft-deleted message
...
  /api/realtime/ephemeral:
    post:
      operationId: publishEphemeral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EphemeralEventRequest"
      responses:
        "202":
          description: Ephemeral event accepted
...
  /api/dms/{conversation_id}:
    get:
      operationId: getDirectConversation
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      responses:
        "200":
          description: Direct conversation
    delete:
      operationId: hideDirectConversation
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      responses:
        "200":
          description: Closed direct conversation for the current user
...
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    channel_id:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
    message_id:
      name: message_id
      in: path
      required: true
      schema:
        type: string
    conversation_id:
      name: conversation_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        template:
          type: string
          enum: [chat, code]
        code_mode:
          type: string
          enum: [single_user, multi_user]
        pull_request_url:
          type: string
          format: uri-reference
          pattern: '^$|^https://github\.com/[^/]+/[^/]+/pull/[1-9][0-9]*$'
          description: Canonical HTTPS github.com pull request URL for this code channel. Send an empty string to clear it.
        pull_request_title:
          type: string
          maxLength: 160
          description: Human-readable title for the primary pull request. Defaults to owner/repo #number.
        archived:
          type: boolean
    UpdateCodeWorkspaceNotesRequest:
      type: object
      minProperties: 1
      properties:
        plan_body:
          type: string
          maxLength: 20000
        goal_body:
          type: string
          maxLength: 20000
...
    CreateMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        body_format:
          type: string
          enum: [markdown]
          default: markdown
        quoted_message_id:
          type: string
          description: |
            Optional ID of a message to quote. The quoted message must live in
            the same channel, DM conversation, or thread as the new message;
            cross-context quoting is rejected with 400. The server captures a
            soft snapshot of the quoted body at send time.
        nonce:
          type: string
          maxLength: 128
          description: |
            Optional client idempotency key for retry-safe message creation.
            Reusing the same nonce with the same body and quote returns the
            existing message with HTTP 200 instead of creating a duplicate.
        topic_id:
          type: string
          description: Optional topic id. Channel-scoped topics can only be used in their channel.
        kind:
          type: string
          enum: [message, agent_commentary, agent_tool]
          description: |
            Durable message kind; omitted values default to message. Agent
            activity kinds require bot-token auth
            with the explicit agent_activity:write scope and are supported on
            channel and direct-message create endpoints.
        turn_id:
          type: string
          description: |
            Optional agent-turn correlation ID. Allowed only with
            agent_commentary or agent_tool; ordinary messages reject it.
...
    UpdateWorkspaceRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
...
    Workspace:
      type: object
      required: [id, route_id, name, slug, icon_url, created_at]
      properties:
        id:
          type: string
        route_id:
          type: string
          description: Immutable public route ID used in app URLs.
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, moderator, member, guest, bot]
...
    EphemeralEventRequest:
      type: object
      required: [workspace_id, type]
      properties:
        workspace_id:
          type: string
        channel_id:
          type: string
        direct_conversation_id:
          type: string
        type:
          type: string
          enum: [typing.started, typing.stopped, presence.changed]
        payload:
          type: object
          additionalProperties: true
...